Atualização

+++ ListaDeExercicios/Calculadora/Calculadora.java
@@ -1,19 +1,19 @@
 public class Calculadora {
 
-    public float Soma(float number1, float number2){
-        float resultado;
+    public int Soma(int number1, int number2){
+        int resultado;
         resultado = number1 + number2;
         return resultado;
     }
 
-    public float Produto(float number1, float number2){
-        float resultado;
+    public int Produto(int number1, int number2){
+        int resultado;
         resultado = number1 * number2;
         return resultado;
     }
 
-    public float Diferenca(float number1, float number2){
-        float resultado;
+    public int Diferenca(int number1, int number2){
+        int resultado;
         resultado = number1 - number2;
         return resultado;
     }
@@ -24,8 +24,8 @@ public class Calculadora {
         return resultado;
     }
 
-    public float Maior(float number1, float number2){
-        float resultado;
+    public int Maior(int number1, int number2){
+        int resultado;
         if(number1 > number2){
             resultado = number1;
         } else {
